refactor: add more stories (WIP)
