+++ .storybook/stories/useSimplification.stories.tsx
@@ -1,10 +1,9 @@
 import * as React from 'react'
+import { withKnobs, number } from '@storybook/addon-knobs'
 
 import { Setup } from '../Setup'
 
-import { useSimplification } from '../../src/useSimplification'
-import { Octahedron } from '../../src/shapes'
-import { withKnobs, number } from '@storybook/addon-knobs'
+import { useSimplification, Octahedron } from '../../src'
 
 export default {
   title: 'Modifiers/useSimplification',
@@ -13,7 +12,7 @@ export default {
 }
 
 function UseSimplificationScene() {
-  const meshRef = useSimplification(number("%", .1, { range: true, max: 1, step: 0.1 }))
+  const meshRef = useSimplification(number('%', 0.1, { range: true, max: 1, step: 0.1 }))
 
   return (
     <Octahedron args={[3, 2]} ref={meshRef}>
